Moved malicious impls into their own directory
