+++ src/main/java/distributed/malicious/MaliciousDistributedAgreement.java
@@ -1,4 +1,4 @@
-package distributed.server.malicious;
+package distributed.malicious;
 
 import distributed.DistributedAgreement;
