+++ src/main/java/distributed/malicious/client/BuggyClient.java
@@ -1,4 +1,4 @@
-package distributed.client.malicious;
+package distributed.malicious.client;
 
 import distributed.client.Client;
 import distributed.server.pojos.Server;
@@ -6,26 +6,23 @@ import distributed.utils.Command;
 import org.apache.log4j.Logger;
 
 import java.util.List;
+import java.util.Random;
 
 
 /**
- * Implementation of a malicious Client. Always lies about the values in the proposal
+ * Implementation of a buggy client that proposes 0 or 1 by chance
  */
-public class MaliciousClient extends Client
+public class BuggyClient extends Client
 {
     private static Logger logger = Logger.getLogger(MaliciousClient.class);
 
     public static void sendValue(List<Server> servers, String value)
     {
-        // Flip the proposed value
-        if(Client.ZERO.equals(value))
-        {
-            value = Client.ONE;
-        }else
-        {
-            value = Client.ZERO;
-        }
-        logger.debug("I'm a bad client so I'm going to propose " + value);
+        // Propose a value by chance
+        Random r = new Random();
+        int randomValue = r.nextInt(2);
+        value = Integer.toString(randomValue);
+        logger.debug("I'm a buggy client so I'm going to propose " + value);
         String cmd = Command.PROPOSE.getCommand() + " " + value + "\n";
         sendCmd(servers, cmd);
     }
